moved way data map to another file
