+++ src/layouts/Tile/index.js
@@ -1,5 +1,4 @@
 import { useSelector } from "react-redux";
-import { selectMovies } from "../../features/movies/movie/moviesSlice";
 import { selectPeople } from "../../features/peoples/people/peopleSlice";
 import { MovieTile } from "./movies/movie";
 import { MovieDetailsTile } from "./movies/movieDetails";
@@ -7,8 +6,7 @@ import { PersonTile } from "./persons/person";
 import { PersonDetailsTile } from "./persons/personDetails";
 import { selectMovieDetails } from "../../features/movies/movieDetails/movieDetailsSlice";
 
-export const Tile = ({ movie, movieDetails, person, role, personDetails }) => {
-  const movies = useSelector(selectMovies);
+export const Tile = ({ movie, movieDetails, person, role, personDetails, key, id, title, release_date, poster_path, genre_ids, vote_average, vote_count }) => {
   const moviesDetails = useSelector(selectMovieDetails);
   const people = useSelector(selectPeople);
 
@@ -16,30 +14,17 @@ export const Tile = ({ movie, movieDetails, person, role, personDetails }) => {
     <>
       {movie ? (
         <>
-          {movies.map(
-            ({
-              key,
-              id,
-              title,
-              release_date,
-              poster_path,
-              genre_ids,
-              vote_average,
-              vote_count,
-            }) => (
-              <MovieTile
-                movie
-                key={key}
-                id={id}
-                title={title}
-                release_date={release_date}
-                poster_path={poster_path}
-                genre_ids={genre_ids}
-                vote_average={vote_average}
-                vote_count={vote_count}
-              />
-            )
-          )}
+          <MovieTile
+            movie
+            key={key}
+            id={id}
+            title={title}
+            release_date={release_date}
+            poster_path={poster_path}
+            genre_ids={genre_ids}
+            vote_average={vote_average}
+            vote_count={vote_count}
+          />
         </>
       ) : (
         ""
